Name Choice System Implemented
On join each player is given 2 alternative name options for 10mins.

These names are cached and marked as used until 10 or more minutes of play time.

They may choose to keep the pre-assigned name or pick from one of the rerolled names.

If they do, they will not be able to reroll again or go back. They are given a confirmation and a welcome msg on join informing them of this new feature.